change Product Callsto be same like in other microservices

+++ src/main/java/pl/edu/pwsztar/domain/dto/BillDto.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -13,6 +15,6 @@ public class BillDto {
     private long userId;
     private String produckts;
     //json -> list off produckts
-    private double cost;
+    private BigDecimal cost;
 
 }
